+++ src/applications/ezr/components/IntroductionPage/SaveInProgressInfo.jsx
@@ -1,18 +1,16 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
-import { connect, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 
 import SaveInProgressIntro from 'platform/forms/save-in-progress/SaveInProgressIntro';
-import { fetchEnrollmentStatus as fetchEnrollmentStatusAction } from '../../utils/actions/enrollment-status';
 import { selectEnrollmentStatus } from '../../utils/selectors/entrollment-status';
 import { selectAuthStatus } from '../../utils/selectors/auth-status';
 import EnrollmentStatusAlert from '../FormAlerts/EnrollmentStatusAlert';
 import VerifiedPrefillAlert from '../FormAlerts/VerifiedPrefillAlert';
 import content from '../../locales/en/content.json';
 
-const SaveInProgressInfo = props => {
-  const { fetchEnrollmentStatus, formConfig, pageList } = props;
-  const { isLoggedOut, isUserLOA3 } = useSelector(selectAuthStatus);
+const SaveInProgressInfo = ({ formConfig, pageList }) => {
+  const { isLoggedOut } = useSelector(selectAuthStatus);
   const { isEnrolledinESR, hasServerError } = useSelector(
     selectEnrollmentStatus,
   );
@@ -22,13 +20,6 @@ const SaveInProgressInfo = props => {
     savedFormMessages,
     customText,
   } = formConfig;
-
-  useEffect(() => {
-    if (isUserLOA3) {
-      fetchEnrollmentStatus();
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
 
   // set the props to use for the SaveInProgressIntro components
   const sipProps = {
@@ -82,16 +73,8 @@ const SaveInProgressInfo = props => {
 };
 
 SaveInProgressInfo.propTypes = {
-  fetchEnrollmentStatus: PropTypes.func,
   formConfig: PropTypes.object,
   pageList: PropTypes.array,
 };
 
-const mapDispatchToProps = {
-  fetchEnrollmentStatus: fetchEnrollmentStatusAction,
-};
-
-export default connect(
-  null,
-  mapDispatchToProps,
-)(SaveInProgressInfo);
+export default SaveInProgressInfo;
