+++ src/applications/ezr/containers/IntroductionPage.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { useSelector } from 'react-redux';
+import { connect, useSelector } from 'react-redux';
 
 import { AUTH_EVENTS } from 'platform/user/authentication/constants';
 import recordEvent from 'platform/monitoring/record-event';
@@ -10,6 +10,7 @@ import {
   externalServices,
 } from 'platform/monitoring/DowntimeNotification';
 
+import { fetchEnrollmentStatus as fetchEnrollmentStatusAction } from '../utils/actions/enrollment-status';
 import { selectEnrollmentStatus } from '../utils/selectors/entrollment-status';
 import { selectAuthStatus } from '../utils/selectors/auth-status';
 import IdentityVerificationAlert from '../components/FormAlerts/IdentityVerificationAlert';
@@ -18,9 +19,9 @@ import SaveInProgressInfo from '../components/IntroductionPage/SaveInProgressInf
 import OMBInfo from '../components/IntroductionPage/OMBInfo';
 import content from '../locales/en/content.json';
 
-const IntroductionPage = ({ route }) => {
+const IntroductionPage = ({ fetchEnrollmentStatus, route }) => {
   const { isLoading } = useSelector(selectEnrollmentStatus);
-  const { isUserLOA1 } = useSelector(selectAuthStatus);
+  const { isUserLOA1, isUserLOA3 } = useSelector(selectAuthStatus);
   const { formConfig, pageList } = route;
   const sipProps = { formConfig, pageList };
 
@@ -28,6 +29,10 @@ const IntroductionPage = ({ route }) => {
 
   useEffect(() => {
     focusElement('.va-nav-breadcrumbs-list');
+    if (isUserLOA3) {
+      fetchEnrollmentStatus();
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return (
@@ -57,7 +62,15 @@ const IntroductionPage = ({ route }) => {
 };
 
 IntroductionPage.propTypes = {
+  fetchEnrollmentStatus: PropTypes.func,
   route: PropTypes.object,
 };
 
-export default IntroductionPage;
+const mapDispatchToProps = {
+  fetchEnrollmentStatus: fetchEnrollmentStatusAction,
+};
+
+export default connect(
+  null,
+  mapDispatchToProps,
+)(IntroductionPage);
